feat: fire handling of get request
